Better Manage of serie's events scope.
On merging with 2.3.2 i have added a 'chart' property to
Highcharts.ux.Serie. This,read only, property is setted by the chart to
point to the chart itself.

```javascript
Ext.define('Chart.ux.Highcharts.Serie', {
    .....
    /**
     * @readonly
     * The {@link Chart.ux.Highcharts} object that has this serie
     *
     * @type Object/Chart.ux.Highcharts
     */
    chart: null,
    ....
```
This can be useful with pointclick event when you need to use an
Ext.Component.

```javascript
pointclick:{
   buffer:10,
   fn:function(serie,point,record,event){
      //Get parent window to replace the chart inside (me)
      var window=this.chart.up('windows');
      ....
   }
}
```
Setting the scope on the listeners at runtime can cause trouble in
ExtJS on parsing the listener